refactor: util 함수로 cursor 설정 폰트설정 분리
